Only prefetch items in the catalog
Updates prefetch to only pull items that are actually present in the
catalog.

Optimizes the augeas file reads to only retrieve items of concern during
a run.

Also ensure that /etc/sysctl.conf is read (if present) regardless of
what target is set in the last resource in the catalog. This ensures
that the global configuration file is always read if present.

Fixes #51